Feature/documentation

+++ crx-web/src/main/java/com/crxmarkets/api/CalculatorResourceImpl.java
@@ -23,6 +23,7 @@ import com.crxmarkets.web.client.shared.CalculatorResource;
 import com.crxmarkets.web.client.shared.HistoryItem;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.regex.Pattern;
 import javax.ejb.EJB;
 import javax.ws.rs.BadRequestException;
 import javax.ws.rs.NotFoundException;
@@ -38,6 +39,13 @@ public class CalculatorResourceImpl implements CalculatorResource {
 
     @EJB
     private HistoryEntityServiceLocal historyEntityService;
+
+    private final Pattern listOfNumbersPattern = Pattern.compile("(([\\-]?\\d+[ ]))+");
+
+    @Override
+    public String testMethod(String data) {
+        return data;
+    }
 
     @Override
     public CalculationResult calculate(CalculationTask task) {
@@ -69,9 +77,23 @@ public class CalculatorResourceImpl implements CalculatorResource {
     }
 
     @Override
-    public long saveInHistory(HistoryItem historyItem) {
-        historyEntityService.create(historyItem);
-        return historyItem.getId();
+    public long saveInHistory(HistoryItem hi) {
+        if (hi == null
+                || hi.getTask() == null
+                || hi.getTask().isEmpty()
+                || hi.getDateTime() == null
+                || hi.getCalculation() == null
+                || hi.getCalculation().isEmpty()) {
+            throw new BadRequestException("History item data is incomplete.");
+        }
+
+//        if (!listOfNumbersPattern.matcher(hi.getCalculation()).matches()
+//                || !listOfNumbersPattern.matcher(hi.getCalculation()).matches()) {
+//            throw new BadRequestException("History item data is incomplete.");
+//        }
+
+        historyEntityService.create(hi);
+        return hi.getId();
     }
 
     @Override
@@ -95,6 +117,10 @@ public class CalculatorResourceImpl implements CalculatorResource {
 
     @Override
     public boolean deleteHistoryItem(long id) {
+        HistoryItem item = historyEntityService.find(id);
+        if (item == null) {
+            throw new NotFoundException();
+        }
         return historyEntityService.delete(id);
     }
